requirements.txt
fastapi==0.117.1
uvicorn==0.37.0
pymongo==4.15.1
twilio==9.8.1
python-dotenv==1.1.1
httpx==0.28.1
pydantic==2.11.9